+++ allensdk/__init__.py
@@ -36,7 +36,7 @@
 import logging
 
 
-__version__ = '1.4.0'
+__version__ = '1.5.0'
 
 
 try: