+++ allensdk/internal/api/behavior_ophys_api.py
@@ -3,11 +3,13 @@ import numpy as np
 import h5py
 import pandas as pd
 import uuid
+from typing import Optional
 import json
 
 from allensdk.api.cache import memoize
 from allensdk.internal.api.ophys_lims_api import OphysLimsApi
-from allensdk.brain_observatory.behavior.sync import get_sync_data, get_stimulus_rebase_function
+from allensdk.brain_observatory.behavior.sync import (
+    get_sync_data, get_stimulus_rebase_function, frame_time_offset)
 from allensdk.brain_observatory.behavior.stimulus_processing import get_stimulus_presentations, get_stimulus_templates, get_stimulus_metadata
 from allensdk.brain_observatory.behavior.metadata_processing import get_task_parameters
 from allensdk.brain_observatory.behavior.running_processing import get_running_df
@@ -19,12 +21,15 @@ from allensdk.internal.api import PostgresQueryMixin
 from allensdk.brain_observatory.behavior.behavior_ophys_api import BehaviorOphysApiBase
 from allensdk.brain_observatory.behavior.trials_processing import get_extended_trials
 from allensdk.internal.core.lims_utilities import safe_system_path
+from allensdk.core.auth_config import LIMS_DB_CREDENTIAL_MAP
+from allensdk.core.authentication import credential_injector, DbCredentials
 
 
 class BehaviorOphysLimsApi(OphysLimsApi, BehaviorOphysApiBase):
 
-    def __init__(self, ophys_experiment_id):
-        super().__init__(ophys_experiment_id)
+    def __init__(self, ophys_experiment_id: int,
+                 lims_credentials: Optional[DbCredentials] = None):
+        super().__init__(ophys_experiment_id, lims_credentials)
 
     @memoize
     def get_sync_data(self):
@@ -59,7 +64,7 @@ class BehaviorOphysLimsApi(OphysLimsApi, BehaviorOphysApiBase):
                 FROM ophys_experiments_visual_behavior_experiment_containers 
                 WHERE ophys_experiment_id= {};
                 '''.format(self.get_ophys_experiment_id())
-        return self.fetchone(query, strict=False)
+        return self.lims_db.fetchone(query, strict=False)
 
     @memoize
     def get_behavior_stimulus_file(self):
@@ -71,7 +76,7 @@ class BehaviorOphysLimsApi(OphysLimsApi, BehaviorOphysApiBase):
                 LEFT JOIN well_known_files stim ON stim.attachable_id=bs.id AND stim.attachable_type = 'BehaviorSession' AND stim.well_known_file_type_id IN (SELECT id FROM well_known_file_types WHERE name = 'StimulusPickle')
                 WHERE oe.id= {};
                 '''.format(self.get_ophys_experiment_id())
-        return safe_system_path(self.fetchone(query, strict=True))
+        return safe_system_path(self.lims_db.fetchone(query, strict=True))
 
     def get_behavior_session_uuid(self):
         behavior_stimulus_file = self.get_behavior_stimulus_file()
@@ -156,9 +161,26 @@ class BehaviorOphysLimsApi(OphysLimsApi, BehaviorOphysApiBase):
         return get_stimulus_templates(data)
 
     @memoize
-    def get_licks(self):
+    def get_sync_licks(self):
         lick_times = self.get_sync_data()['lick_times']
         return pd.DataFrame({'time': lick_times})
+
+    @memoize
+    def get_licks(self):
+        behavior_stimulus_file = self.get_behavior_stimulus_file()
+        data = pd.read_pickle(behavior_stimulus_file)
+        rebase_function = self.get_stimulus_rebase_function()
+        # Get licks from pickle file (need to add an offset to align with
+        # the trial_log time stream)
+        lick_frames = (data["items"]["behavior"]["lick_sensors"][0]
+                       ["lick_events"])
+        vsyncs = data["items"]["behavior"]["intervalsms"]
+        vsync_times_raw = np.hstack((0, vsyncs)).cumsum() / 1000.0  # cumulative time
+        vsync_offset = frame_time_offset(data)
+        vsync_times = vsync_times_raw + vsync_offset
+        lick_times = [vsync_times[frame] for frame in lick_frames]
+        # Align pickle data with sync time stream
+        return pd.DataFrame({"time": list(map(rebase_function, lick_times))})
 
     @memoize
     def get_rewards(self):
@@ -229,7 +251,7 @@ class BehaviorOphysLimsApi(OphysLimsApi, BehaviorOphysApiBase):
                 LEFT JOIN well_known_files wkf ON wkf.attachable_id=oe.id AND wkf.well_known_file_type_id IN (SELECT id FROM well_known_file_types WHERE name = 'BehaviorOphysNwb')
                 WHERE oe.id = {};
                 '''.format(self.get_ophys_experiment_id())
-        return safe_system_path(self.fetchone(query, strict=True))
+        return safe_system_path(self.lims_db.fetchone(query, strict=True))
 
     def get_stimulus_rebase_function(self):
         stimulus_timestamps_no_monitor_delay = self.get_sync_data()['stimulus_times_no_delay']
@@ -246,7 +268,8 @@ class BehaviorOphysLimsApi(OphysLimsApi, BehaviorOphysApiBase):
     @staticmethod
     def get_ophys_experiment_df():
 
-        api = PostgresQueryMixin()
+        api = (credential_injector(LIMS_DB_CREDENTIAL_MAP)
+               (PostgresQueryMixin)())
         query = '''
                 SELECT
 
@@ -274,7 +297,8 @@ class BehaviorOphysLimsApi(OphysLimsApi, BehaviorOphysApiBase):
     @staticmethod
     def get_containers_df(only_passed=True):
 
-        api = PostgresQueryMixin()
+        api = (credential_injector(LIMS_DB_CREDENTIAL_MAP)
+               (PostgresQueryMixin)())
         if only_passed is True:
             query = '''
                     SELECT *